refactor text link tests/driver
