+++ src/Viewer/TextLink/TextLink.spec.js
@@ -23,63 +23,6 @@ describe('TextLink', () => {
     expect(driver.getContent()).toBe(children);
   });
 
-  it('should be with medium size by defualt', () => {
-    const driver = createDriver(<TextLink link=""/>);
-    expect(driver.getSize()).toBe('medium');
-  });
-
-  it('should be with small size', () => {
-    const driver = createDriver(<TextLink link="" size="small"/>);
-    expect(driver.getSize()).toBe('small');
-  });
-
-  it('should be with dark background', () => {
-    const driver = createDriver(<TextLink link="" darkBackground size="small"/>);
-    expect(driver.isDarkBackground()).toBeTruthy();
-  });
-
-  it('should be with light background', () => {
-    const driver = createDriver(<TextLink link="" size="small"/>);
-    expect(driver.isLightBackground()).toBeTruthy();
-  });
-
-  it('should be with underline', () => {
-    const driver = createDriver(<TextLink link="" underlineStyle="always"/>);
-    expect(driver.isUnderline()).toBeTruthy();
-  });
-
-  it('should not be with underline', () => {
-    const driver = createDriver(<TextLink link="" underlineStyle="never"/>);
-    driver.hover();
-    expect(driver.isUnderline()).toBeFalsy();
-  });
-
-  it('should not be with underline by defualt', () => {
-    const driver = createDriver(<TextLink link=""/>);
-    expect(driver.isUnderline()).toBeFalsy();
-  });
-
-  it('should have underline on hover', () => {
-    const driver = createDriver(<TextLink link=""/>);
-    driver.hover();
-    expect(driver.isUnderline()).toBeTruthy();
-  });
-
-  it('should have a link', () => {
-    const driver = createDriver(<TextLink link="https://www.wix.com"/>);
-    expect(driver.getLink()).toBe('https://www.wix.com/');
-  });
-
-  it('should have a rel', () => {
-    const driver = createDriver(<TextLink rel="bookmark" target="_blank" link="https://www.wix.com"/>);
-    expect(driver.getRel()).toBe('bookmark');
-  });
-
-  it('should have a target', () => {
-    const driver = createDriver(<TextLink target="_blank" link="https://www.wix.com"/>);
-    expect(driver.getTarget()).toBe('_blank');
-  });
-
   it('should return the default color', () => {
     const defaultColor = 'rgb(24, 210, 222)';
     const driver = createDriver(<TextLink link="https://www.wix.com"/>);
